Crate nbt

MC Named Binary Tag type.

Modules

serialize
Contains functions for serializing arbitrary objects to the Named Binary Tag format.

Structs

Blob
A generic, complete object in Named Binary Tag format.

Enums

Error
Errors that may be encountered when constructing, parsing, or encoding NbtValue and NbtBlob objects.

Value
Values which can be represented in the Named Binary Tag format.

Type Definitions

Result
A convenient alias type for results when reading/writing the Named Binary Tag format.