Module nbt::serialize

Contains functions for serializing arbitrary objects to the Named Binary Tag format.

For working with existing serialization implementations, see to_writer. For custom types, implement the NbtFmt trait.

Traits

NbtFmt
A trait indicating that the type has a Named Binary Tag representation.

Functions

close_nbt
A convenience function for closing NBT format objects.

emit_next_header
Extracts the next header (tag and name) from an NBT format source.

from_reader
Deserializes an object in NBT format from a given source.

read_bare_nbt

to_writer
Serializes an object into NBT format at a given destination.