Module nbt::serialize::raw

Contains raw, primitive functions for serializing and deserializing basic Named Binary Tag types.

This submodule is not intended for general use, but is exposed for those interested in writing fast NBT encoding/decoding by hand, where it may be quite useful.

For the higher-level abstraction over these primitives, see the NbtFmt trait in the parent module.

Functions

read_bare_byte

read_bare_byte_array

read_bare_double

read_bare_float

read_bare_int

read_bare_int_array

read_bare_list

read_bare_long

read_bare_short

read_bare_string

write_bare_byte

write_bare_byte_array

write_bare_compound

write_bare_double

write_bare_float

write_bare_int

write_bare_int_array

write_bare_list

write_bare_long

write_bare_short

write_bare_string