Struct nbt::NbtBlob

pub struct NbtBlob {
    // some fields omitted
}

An object in the Named Binary Tag (NBT) file format.

This is essentially a map of names to NbtValues, with an optional top- level name of its own. It can be created in a similar way to a HashMap, or read from an io::Read source, and its binary representation can be written to an io::Write destination.

These read and write methods support both uncompressed and compressed (through Gzip or zlib compression) methods.

use nbt::{NbtBlob, NbtValue};

// Create a `NbtBlob` from key/value pairs.
let mut nbt = NbtBlob::new("".to_string());
nbt.insert("name".to_string(), "Herobrine").unwrap();
nbt.insert("health".to_string(), 100i8).unwrap();
nbt.insert("food".to_string(), 20.0f32).unwrap();

// Write a compressed binary representation to a byte array.
let mut dst = Vec::new();
nbt.write_zlib(&mut dst).unwrap();

Methods

impl NbtBlob

fn new(title: String) -> NbtBlob

Create a new NBT file format representation with the given name.

fn from_reader(src: &mut Read) -> Result<NbtBlob>

Extracts an NbtBlob object from an io::Read source.

fn from_gzip(src: &mut Read) -> Result<NbtBlob>

Extracts an NbtBlob object from an io::Read source that is compressed using the Gzip format.

fn from_zlib(src: &mut Read) -> Result<NbtBlob>

Extracts an NbtBlob object from an io::Read source that is compressed using the zlib format.

fn write(&self, dst: &mut Write) -> Result<()>

Writes the binary representation of this NbtBlob to an io::Write destination.

fn write_gzip(&self, dst: &mut Write) -> Result<()>

Writes the binary representation of this NbtBlob, compressed using the Gzip format, to an io::Write destination.

fn write_zlib(&self, dst: &mut Write) -> Result<()>

Writes the binary representation of this NbtBlob, compressed using the Zlib format, to an io::Write dst.

fn insert<V>(&mut self, name: String, value: V) -> Result<()> where V: Into<NbtValue>

Insert an NbtValue with a given name into this NbtBlob object. This method is just a thin wrapper around the underlying HashMap method of the same name.

This method will also return an error if a NbtValue::List with heterogeneous elements is passed in, because this is illegal in the NBT file format.

fn len(&self) -> usize

The uncompressed length of this NbtBlob, in bytes.