Trait nbt::serialize::NbtFmt

pub trait NbtFmt {
    fn to_bare_nbt<W>(&self, dst: &mut W) -> Result<()> where W: Write;

    fn to_nbt<W, S>(&self, dst: &mut W, name: S) -> Result<()> where W: Write, S: AsRef<str> { ... }
    fn tag() -> u8 { ... }
    fn is_bare() -> bool { ... }
}

A trait indicating that the type has a Named Binary Tag representation.

Keep in mind that not all Rust types (notably unsigned integers) have an obvious NBT representation, and so structs that implement this trait may have to convert them to one that does.

Serialization

To enable NBT serialization of a type, implement the to_bare_nbt method. Many basic types (i8, i16, String, etc.) implement NbtFmt, enabling the use of to_nbt when writing this implementation. The close_nbt function is also provided to aid in implementing custom serialization, given that types will likely advertise themselves as Compound-like. For example:

extern crate nbt;

use nbt::serialize::{NbtFmt, to_writer, close_nbt};

struct MyMob {
    name: String,
    health: i8
}

impl NbtFmt for MyMob {
    fn to_bare_nbt<W>(&self, dst: &mut W) -> nbt::Result<()>
       where W: std::io::Write
    {
        try!(self.name.to_nbt(dst, "name"));
        try!(self.health.to_nbt(dst, "health"));

        close_nbt(dst)
    }
}

fn main() {
    let mut bytes = Vec::new();
    let mob = MyMob { name: "Dr. Evil".to_string(), health: 240 };

    to_writer(&mut bytes, mob).unwrap();
}

Required Methods

fn to_bare_nbt<W>(&self, dst: &mut W) -> Result<()> where W: Write

Convert this type to NBT format using the specified io::Write destination, but does not serialize its identifying NBT tag or name.

Provided Methods

fn to_nbt<W, S>(&self, dst: &mut W, name: S) -> Result<()> where W: Write, S: AsRef<str>

Convert this type to NBT format using the specified io::Write destination, incuding its tag and a given name.

fn tag() -> u8

Indicates the NBT tag that this type corresponds to. Most custom types (usually structs) will advertise the default, 0x0a, which is the default.

fn is_bare() -> bool

Indicates whether this type is "bare", in that it must be wrapped in an NBT Compound before serialization. By default this is false, since most imeplementations will be Compound-like objects. Primitive NBT types (i8, i16, String, etc.) return true.